allow async method to not return promise
Prior to this commit, async methods had to strictly return a promise.
This needed to be improved in order to be able to return a simple value
in an async method, just like in TypeScript.

Since this is now allowed, this is why there is a new await method which
does not take a Promise parameter, but a generic type. (this
modification was in previous commit:
2b784364157e54903b72d1dc0463b4672d140b20)