Add support for transpiling collections to ES6
This commit adds an Adapter that transpiles sets using ES6 Set
implementation.
The goal of the adapter is to use a standard API for the transpiled
collections in order to abstract implementation details of collections
as well as enable
inter-operability between collections just as in Java.
For example in Java most collections can be initialized with any
arbitrary collection and never need to know how that collection is
implemented.

Having the transpiled code work only via the API at collection
boundaries enables this.
However this means that collections transpiled by this adapter are not
compatible with those transpiled by other adapters - this should be less
of an issue once the adapter has support to handle all collection types.

Related #548
Fixes #549