fix #618 : element and type associated with tree could sometimes not be retrieved
In some side cases, a tree's CompilationUnit is not the current scanned
compilation unit, and it is hard to retrieved corresponding Element or
TypeMirror, because Javac Tree API requires a TreePath to get those
elements, and TreePath can only be constructed from a CompilationUnit.

This commit introduces a hack, encapsulated in Util.java, to read internal
sym and type fields from JC* trees implementations as a fallback.