Fixed issue #730
When an overloaded constructor calls another one using "this(...)", the `printInlinedMethod()` function is called
recursively, which initializes the object fields.  In this scenario, we cannot re-initialized those fields after that call,
since we will lose any changes due to constructor parameters passed to `this(...)`.

The fix is completely localized and simple, relying on the existing boolean `skipFirst` variable already set for
the reentrant scenario.