Add support @TsNoCheck annotation, adds "// @ts-nocheck" to TypeScript
Typescript would complain about certain errors in the generated file. In
some newer versions of TypeScript, an assert that is always true would
not be permitted.

We currently have two such asserts in our tests, which prevents us from
upgrading the TypeScript version.

Add support for a Java @TsNoCheck annotation, which controls the
inclusion of the corresponding TypeScript comment. While there is no
mandatory package for the annotation, add it to jsweet.lang in core-lib.

Fix tests to use the new annotation.