Add /opt/local/bin to the Eclipse path workaround
In some environments, "node" may not be on the PATH.

Add another path that is commonly supported, /opt/local/bin.

Note that we cannot add /opt/homebrew/bin since there's a dependency
that would need changing as well (typescript.java).

For the time being, it is encouraged to add a symbolic link from
/opt/homebrew/bin/node to /opt/local/bin/node (and npm) if that's
necessary.